+++ mlir/lib/ExecutionEngine/AxiUtilsMock.cpp
@@ -19,11 +19,11 @@
 // AXI_APIV1
 // =============================================================================
 
-extern "C" void dma_init(unsigned int dma_address,
-                         unsigned int dma_input_address,
-                         unsigned int dma_input_buffer_size,
-                         unsigned int dma_output_address,
-                         unsigned int dma_output_buffer_size) {
+extern "C" void dma_init(uint64_t dma_address,
+                         uint64_t dma_input_address,
+                         uint64_t dma_input_buffer_size,
+                         uint64_t dma_output_address,
+                         uint64_t dma_output_buffer_size) {
   std::cout << "Called: " << __func__ << std::endl;
   std::cout << "\t" << dma_address << std::endl;
   std::cout << "\t" << dma_input_address << std::endl;
@@ -36,63 +36,63 @@ extern "C" void dma_init(unsigned int dma_address,
 
 extern "C" void dma_free() { std::cout << "Called: " << __func__ << std::endl; }
 
-extern "C" unsigned int *dma_get_inbuffer() {
+extern "C" uint64_t *dma_get_inbuffer() {
   std::cout << "Called: " << __func__ << std::endl;
   return 0;
 }
 
-extern "C" unsigned int *dma_get_outbuffer() {
+extern "C" uint64_t *dma_get_outbuffer() {
   std::cout << "Called: " << __func__ << std::endl;
   return 0;
 }
 
-extern "C" int dma_copy_to_inbuffer(unsigned int *host_src_address,
-                                    int data_length, int offset) {
+extern "C" int64_t dma_copy_to_inbuffer(uint64_t *host_src_address,
+                                    int64_t data_length, int64_t offset) {
   std::cout << "Called: " << __func__ << std::endl;
   return 0;
 }
 
-extern "C" int dma_copy_from_outbuffer(unsigned int *host_dst_address,
-                                       int data_length, int offset) {
+extern "C" int64_t dma_copy_from_outbuffer(uint64_t *host_dst_address,
+                                       int64_t data_length, int64_t offset) {
   std::cout << "Called: " << __func__ << std::endl;
   return 0;
 }
 
 template <typename T>
-int mlir_dma_copy_to_inbuffer(const DynamicMemRefType<T> &src, int data_length,
-                              int offset) {
+int64_t mlir_dma_copy_to_inbuffer(const DynamicMemRefType<T> &src, int64_t data_length,
+                              int64_t offset) {
   std::cout << "Called: " << __func__ << std::endl;
   return 0;
 }
 
-extern "C" int _mlir_ciface_copy_to_inbuffer_f32(UnrankedMemRefType<float> *M,
+extern "C" int64_t _mlir_ciface_copy_to_inbuffer_f32(UnrankedMemRefType<float> *M,
                                                  int64_t offset) {
   mlir_dma_copy_to_inbuffer(DynamicMemRefType<float>(*M), 0, offset);
   return 0;
 }
 
-extern "C" int copy_to_inbuffer_f32(int64_t rank, void *ptr, int64_t offset) {
+extern "C" int64_t copy_to_inbuffer_f32(int64_t rank, void *ptr, int64_t offset) {
   UnrankedMemRefType<float> descriptor = {rank, ptr};
   return 0;
 }
 
-extern "C" int copy_to_inbuffer_i32(int64_t rank, void *ptr, int64_t offset) {
-  UnrankedMemRefType<int> descriptor = {rank, ptr};
+extern "C" int64_t copy_to_inbuffer_i32(int64_t rank, void *ptr, int64_t offset) {
+  UnrankedMemRefType<int64_t> descriptor = {rank, ptr};
   return 0;
 }
 
-extern "C" int copy_from_outbuffer_f32(int64_t rank, void *ptr,
+extern "C" int64_t copy_from_outbuffer_f32(int64_t rank, void *ptr,
                                        int64_t offset) {
   UnrankedMemRefType<float> descriptor = {rank, ptr};
   return 0;
 }
 
-extern "C" int dma_start_send(int length, int offset) {
+extern "C" int64_t dma_start_send(int64_t length, int64_t offset) {
   std::cout << "Called: " << __func__ << std::endl;
   return 0;
 }
 
-extern "C" int dma_check_send() {
+extern "C" int64_t dma_check_send() {
   std::cout << "Called: " << __func__ << std::endl;
   return 0;
 }
@@ -101,7 +101,7 @@ extern "C" void dma_wait_send() {
   std::cout << "Called: " << __func__ << std::endl;
 }
 
-extern "C" int dma_start_recv(int length, int offset) {
+extern "C" int64_t dma_start_recv(int64_t length, int64_t offset) {
   std::cout << "Called: " << __func__ << std::endl;
   return 0;
 }
@@ -110,18 +110,18 @@ extern "C" void dma_wait_recv() {
   std::cout << "Called: " << __func__ << std::endl;
 }
 
-extern "C" int dma_check_recv() {
+extern "C" int64_t dma_check_recv() {
   std::cout << "Called: " << __func__ << std::endl;
   return 0;
 }
 
-extern "C" unsigned int dma_set(unsigned int *dma_virtual_address, int offset,
-                                unsigned int value) {
+extern "C" uint64_t dma_set(uint64_t *dma_virtual_address, int64_t offset,
+                                uint64_t value) {
   std::cout << "Called: " << __func__ << std::endl;
   return 0;
 }
 
-extern "C" unsigned int dma_get(unsigned int *dma_virtual_address, int offset) {
+extern "C" uint64_t dma_get(uint64_t *dma_virtual_address, int64_t offset) {
   std::cout << "Called: " << __func__ << std::endl;
   return 0;
 }
